feat: add LogTab component
